+++ www/views/Extjs3/javascript/data/JsonStore.js
@@ -90,19 +90,16 @@ GO.data.JsonStore = function(config) {
 
 	
 	this.on('exception',		
-		function( store, type, action, options, response, arg){
+		function( store, type, action, options, response){
 
-			console.error(arg, response);
 
 			if(response.isAbort) {
 				//ignore aborts.
 			} else if(response.isTimeout){
+				console.error(response);
+
 				GO.errorDialog.show(t("The request timed out. The server took too long to respond. Please try again."));
-			}else	if(response.status==0)
-			{
-				//silently ignore because auto refreshing jobs often get here somehow??
-				//GO.errorDialog.show(t("Could not connect to the server. Please check your internet connection."), "");
-			}else if(!this.reader.jsonData || GO.jsonAuthHandler(this.reader.jsonData, this.load, this))
+			}else	if(!this.reader.jsonData || GO.jsonAuthHandler(this.reader.jsonData, this.load, this))
 			{
 				var msg;
 
@@ -118,6 +115,11 @@ GO.data.JsonStore = function(config) {
 						GO.errorDialog.show(msg);
 					}
 				}					
+			}else
+			{
+				console.error(response);
+
+				GO.errorDialog.show(t("Failed to send the request to the server. Please check your internet connection."));
 			}
 		}
 		,this);