+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Core: Improved connection error handling
+
 28-02-2020 6.4.123
 - Calendar: Fixed error when saving resource
 - Core: Fixed error on saving groups